Added Tailwind CSS
Not sure I like this. The HTML becomes very verbose and feels wrong to me. We are explaining what each element should look like, in the HTML. It’s like say class = “red” instead of class = “error”… seems backwards… but I guess it is “fast” in that you don’t need to go to the CSS file to add anything but I’m not sure this is really a good idea. Anyway, the tutorial is using Tailwind so I’m going with the flow.

Signed-off-by: Aaron Rosenzweig <aaron@chatnbike.com>